Version 1.4, ya acabado. Falta prueba con profiler y actualizar diagramas.
