+++ HOJA2/UMLHOJA2.ucls
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<class-diagram version="1.1.8" icons="true" always-add-relationships="false" generalizations="true" realizations="true" 
+<class-diagram version="1.1.10" icons="true" always-add-relationships="false" generalizations="true" realizations="true" 
   associations="true" dependencies="false" nesting-relationships="true">  
-  <class id="1" language="java" name="main" project="HOJA2" file="/HOJA2/src/main.java" binary="false" 
-    corner="BOTTOM_RIGHT">    
+  <class id="1" language="java" name="MainCalculadora" project="HOJA2" file="/HOJA2/src/MainCalculadora.java" 
+    binary="false" corner="BOTTOM_RIGHT">    
     <position height="153" width="172" x="196" y="37"/>    
     <display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
       sort-features="false" accessors="true" visibility="true">      
@@ -19,16 +19,7 @@
       <operations public="true" package="true" protected="true" private="true" static="true"/>    
     </display>  
   </interface>  
-  <class id="3" language="java" name="StackVector" project="HOJA2" file="/HOJA2/src/StackVector.java" binary="false" 
-    corner="BOTTOM_RIGHT">    
-    <position height="189" width="115" x="41" y="37"/>    
-    <display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
-      sort-features="false" accessors="true" visibility="true">      
-      <attributes public="true" package="true" protected="true" private="true" static="true"/>      
-      <operations public="true" package="true" protected="true" private="true" static="true"/>    
-    </display>  
-  </class>  
-  <class id="4" language="java" name="Pruebas" project="HOJA2" file="/HOJA2/src/Pruebas.java" binary="false" 
+  <class id="3" language="java" name="Pruebas" project="HOJA2" file="/HOJA2/src/Pruebas.java" binary="false" 
     corner="BOTTOM_RIGHT">    
     <position height="99" width="87" x="408" y="37"/>    
     <display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
@@ -37,18 +28,18 @@
       <operations public="true" package="true" protected="true" private="true" static="true"/>    
     </display>  
   </class>  
-  <association id="5">    
+  <association id="4">    
     <end type="SOURCE" refId="1" navigable="false">      
-      <attribute id="6" name="miStack"/>      
-      <multiplicity id="7" minimum="0" maximum="1"/>    
+      <attribute id="5" name="miStack">        
+        <position height="0" width="0" x="0" y="0"/>      
+      </attribute>      
+      <multiplicity id="6" minimum="0" maximum="1">        
+        <position height="0" width="0" x="0" y="0"/>      
+      </multiplicity>    
     </end>    
     <end type="TARGET" refId="2" navigable="true"/>    
     <display labels="true" multiplicity="true"/>  
   </association>  
-  <realization id="8">    
-    <end type="SOURCE" refId="3"/>    
-    <end type="TARGET" refId="2"/>  
-  </realization>  
   <classifier-display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
     sort-features="false" accessors="true" visibility="true">    
     <attributes public="true" package="true" protected="true" private="true" static="true"/>
