--- HOJA2/src/Pruebas.java
@@ -1,13 +1,0 @@
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-
-public class Pruebas {
-
-	@Test
-	public void test() {
-		fail("Not yet implemented");
-	}
-
-}
