+++ HOJA2/src/Stack.java
@@ -1,3 +1,8 @@
+/**
+*Interfaz para implementar una estructura de Pila (Stack). 
+*@author UVG 
+*@version 1.4 Julio 24, 2016
+*/
 public interface Stack<E> 
 {
